Better organize audio classes
